+++ lib/persistence/dao/talent_dao.dart
